+++ packages/types/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./apiResponse";
 export * from "./auth";
 export * from "./common";
+export * from "./httpHeaders";
 export * from "./notification";
 export * from "./otp";
 export * from "./platform";